Additional clean-up
