+++ Assets/Scripts/FiniteStack.cs.meta
